feat: calculate timelines of all beams
To be completely honest, I didn't come up with this on my own.
I had to look at the visualisation and saw a pattern I thought:
"Wow, this might work, lemme think about this".

So yeah, not completely original solution but I also didn't copy code,
only saw the "visualisation".

With this, part 2 is completed